Added @Schedule annotated EJB for reporting the number of indians
- added more challenges and solutions

- fixed a weird auto boxing problem leading to strange NPEs upon entity
creation.

- fixed bug in GatheringTribe copy constructor where it didn't set the
id, leading to TransientPropertyValueException Hibernate exceptions.

- introdced dependency management of local dependencies

+++ moccasin-webapp/src/main/java/net/skybert/moccasin/ejb/IndianService.java
@@ -14,7 +14,7 @@ public interface IndianService
 
   public List<Tribe> allTribes();
 
-  public long create(Indian indian);
+  public Integer create(Indian indian);
 
   public Object findTribe(Integer valueOf);
